added processing of accounts + created structure of website + fixes

+++ MeetingRoomsServiceApp/Controllers/UsersController.cs
@@ -8,28 +8,41 @@ using Microsoft.EntityFrameworkCore;
 using DataAccessLayer;
 using DataAccessLayer.Models;
 using BusinessLogic.Services;
+using System.Web.Helpers;
+using MeetingRoomsServiceApp.Models;
+using System.Web;
+using System.IO;
+using Microsoft.AspNetCore.Http;
+using AutoMapper;
 
 namespace MeetingRoomsServiceApp.Controllers
 {
     public class UsersController : Controller
     {
         private readonly UserService _service;
+        private readonly ReservationService _reservationService;
+        private readonly MapperConfiguration editIndexConfig = new(cfg => cfg
+        .CreateMap<User, UserEditModel>()
+        .ForMember(x => x.Image, opt => opt.Ignore())
+        .ForMember(x => x.EditNotValid, opt => opt.Ignore())
+        .ForMember(x => x.UserId, opt => opt.Ignore()));
+        private readonly MapperConfiguration editConfig = new(cfg => cfg
+        .CreateMap<UserEditModel, User>()
+        .ForMember(x => x.Image, opt => opt.Ignore()));
+        private readonly MapperConfiguration config = new(cfg => cfg
+        .CreateMap<UserPostModel, User>()
+        .ForMember("Image", opt => opt.Ignore()));
 
-        public UsersController(UserService service)
+        public UsersController(UserService service, ReservationService reservationService)
         {
             _service = service;
+            _reservationService = reservationService;
         }
 
         // GET: Users
         public async Task<IActionResult> Index()
         {
             return View(await _service.GetAll());
-        }
-
-        // GET: Users/Details/5
-        public async Task<IActionResult> Details(int id)
-        {
-            return View(await _service.GetById(id));
         }
 
         // GET: Users/Create
@@ -43,21 +56,39 @@ namespace MeetingRoomsServiceApp.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Id,Login,Password,Name")] User user)
+        public async Task<IActionResult> Create([Bind("Id,Login,Password,Name,Image")] UserPostModel userPost)
         {
             if (ModelState.IsValid)
             {
+                var mapper = new Mapper(config);
+                var user = mapper.Map<UserPostModel, User>(userPost);
+                if (userPost.Image != null)
+                {
+                    var stream = new MemoryStream();
+                    await userPost.Image.CopyToAsync(stream);
+                    user.Image = stream.ToArray();
+                }
+                
                 await _service.Add(user);
-                return RedirectToAction(nameof(Index));
+                return RedirectToAction(nameof(Login));
             }
-            return View(user);
+            return View(userPost);
         }
 
         // GET: Users/Edit/5
-        public async Task<IActionResult> Edit(int id)
+        public async Task<IActionResult> Edit(int id, int userId)
         {
             var user = await _service.GetById(id);
-            return View(user);
+            var mapper = new Mapper(editIndexConfig);
+            var userPost = mapper.Map<User, UserEditModel>(user);
+            //var userPost = new UserPostModel()
+            //{
+            //    Id = user.Id,
+            //    Login = user.Login,
+            //    Password = user.Password,
+            //    Name = user.Name
+            //};
+            return View(userPost);
         }
 
         // POST: Users/Edit/5
@@ -65,14 +96,26 @@ namespace MeetingRoomsServiceApp.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("Id,Login,Password,Name")] User user)
+        public async Task<IActionResult> Edit(int id, int userId, [Bind("Id,Login,Password,Name,Image")] UserEditModel userEdit)
         {
+            if (id != userId) return View(new UserEditModel()
+            {
+                EditNotValid = true
+            });
+            var mapper = new Mapper(editConfig);
+            var user = mapper.Map<UserEditModel, User>(userEdit);
+            if (userEdit.Image != null)
+            {
+                var stream = new MemoryStream();
+                await userEdit.Image.CopyToAsync(stream);
+                user.Image = stream.ToArray();
+            }
             await _service.Update(user);
-            return View(user);
+            return RedirectToAction("IndexWithDates", "Reservations");
         }
 
         // GET: Users/Delete/5
-        public async Task<IActionResult> Delete(int id)
+        public async Task<IActionResult> Delete(int id, int userId)
         {
             var user = await _service.GetById(id);
             return View(user);
@@ -81,10 +124,57 @@ namespace MeetingRoomsServiceApp.Controllers
         // POST: Users/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
+        public async Task<IActionResult> DeleteConfirmed(int id, int userId)
         {
+            if (id != userId) return View(new UserDeleteModel()
+            {
+                DeleteNotValid = true
+            });
             await _service.Delete(id);
             return RedirectToAction(nameof(Index));
+        }
+
+        public async Task<IActionResult> Login()
+        {
+            return View();
+        }
+
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> Login(UserLoginModel user)
+        {
+            if (await _service.Login(user.Login, user.Password))
+            {
+                var commonUser = await _service.GetByLogin(user.Login);
+                UserCacheModel userCache = new()
+                {
+                    Id = commonUser.Id,
+                    Image = commonUser.Image
+                };
+                WebCache.Set("LoggedIn", userCache, 60, true);
+                return RedirectToAction("IndexWithDates", "Reservations");
+            }
+            else return View(new UserLoginModel()
+            {
+                InfoNotValid = true
+            });
+        }
+
+        public async Task<IActionResult> Logout()
+        {
+            WebCache.Remove("LoggedIn");
+            return RedirectToAction("IndexWithDates", "Reservations");
+        }
+
+        public async Task<IActionResult> UserPage()
+        {
+            UserCacheModel user = WebCache.Get("LoggedIn");
+            return View(new UserPageModel()
+            {
+                Id = user.Id,
+                Reservations = _reservationService.GetForUser(user.Id),
+                User = await _service.GetById(user.Id)
+            });
         }
     }
 }
